requirements.txt
streamlit
pymongo
python-dotenv
langchain
langchain-openai